01-07 合同备份/22-04-07 赶集找铺开发/项目代码/小程序后端/DRF后端V2/polls/serializers.py
from pyexpat import model
from turtle import update
from django.forms import ModelChoiceField
from rest_framework import serializers
from polls.models import *

class recommentUserSerializer(serializers.ModelSerializer):
    class Meta:
        # 与test表对应
        model = recommentUser
        # 取全部字段
        fields = '__all__'
        # 级联查询
        depath = 1

