01-07 合同备份/22-04-07 赶集找铺开发/项目代码/小程序后端/DRF后端V2/requirements.txt
Django==3.2.12
djangorestframework==3.13.1
PyMySQL==1.0.2
django-sslserver
django-filter
anyio
matplotlib